+++ style_bert_vits2/nlp/japanese/normalizer.py
@@ -121,7 +121,7 @@ def normalize_text(text: str) -> str:
     # 結合文字の濁点・半濁点を削除
     # 通常の「ば」等はそのままのこされる、「あ゛」は上で「あ゙」になりここで「あ」になる
     res = res.replace("\u3099", "")  # 結合文字の濁点を削除、る゙ → る
-    res = res.replace("\u309A", "")  # 結合文字の半濁点を削除、な゚ → な
+    res = res.replace("\u309a", "")  # 結合文字の半濁点を削除、な゚ → な
     return res